requirements.txt
# requirements.txt

selenium
pymongo
google-cloud-documentai
google-generativeai
pyyaml
python-dotenv
requests